pyproject.toml
[project]
name = "ai-mcp-server-langchain"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "altair==5.5.0",
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "attrs==25.3.0",
    "blinker==1.9.0",
    "boto3==1.39.9",
    "botocore==1.39.9",
    "cachetools==6.1.0",
    "certifi==2025.7.14",
    "charset-normalizer==3.4.2",
    "click==8.2.1",
    "fastapi>=0.116.1",
    "gitdb==4.0.12",
    "gitpython==3.1.44",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "httpx-sse==0.4.1",
    "idna==3.10",
    "ipykernel>=6.30.0",
    "jinja2==3.1.6",
    "jmespath==1.0.1",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "jsonschema==4.25.0",
    "jsonschema-specifications==2025.4.1",
    "langchain>=0.3.26",
    "langchain-aws==0.2.28",
    "langchain-community>=0.3.27",
    "langchain-core==0.3.69",
    "langchain-mcp-adapters==0.1.9",
    "langchain-ollama==0.3.5",
    "langgraph==0.5.3",
    "langgraph-checkpoint==2.1.1",
    "langgraph-prebuilt==0.5.2",
    "langgraph-sdk==0.1.73",
    "langsmith==0.4.8",
    "markupsafe==3.0.2",
    "mcp[cli]==1.12.0",
    "narwhals==1.47.1",
    "numpy==2.3.1",
    "ollama==0.5.1",
    "orjson==3.11.0",
    "ormsgpack==1.10.0",
    "packaging==25.0",
    "pandas==2.3.1",
    "pillow==11.3.0",
    "protobuf==6.31.1",
    "pyarrow==21.0.0",
    "pydantic==2.11.7",
    "pydantic-core==2.33.2",
    "pydantic-settings==2.10.1",
    "pydeck==0.9.1",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.1",
    "python-multipart==0.0.20",
    "pytz==2025.2",
    "pyyaml==6.0.2",
    "referencing==0.36.2",
    "requests==2.32.4",
    "requests-toolbelt==1.0.0",
    "rpds-py==0.26.0",
    "s3transfer==0.13.1",
    "six==1.17.0",
    "smmap==5.0.2",
    "sniffio==1.3.1",
    "sse-starlette==2.4.1",
    "starlette==0.47.2",
    "streamlit==1.47.0",
    "tenacity==9.1.2",
    "toml==0.10.2",
    "tornado==6.5.1",
    "typing-extensions==4.14.1",
    "typing-inspection==0.4.1",
    "tzdata==2025.2",
    "urllib3==2.5.0",
    "uvicorn==0.35.0",
    "xxhash==3.5.0",
    "zstandard==0.23.0",
]
